+++ src/interrupts/isr.h
@@ -21,10 +21,10 @@
 #define IRQ15 47
 
 typedef struct registers {
-    uint32 ds;                  
-    uint32 edi, esi, ebp, esp, ebx, edx, ecx, eax; 
-    uint32 int_no, err_code;    
-    uint32 eip, cs, eflags, useresp, ss; 
+    size_t ds;                  
+    size_t edi, esi, ebp, esp, ebx, edx, ecx, eax; 
+    size_t int_no, err_code;    
+    size_t eip, cs, eflags, useresp, ss; 
 } registers_t;
 
 typedef void (*isr_t)(registers_t *);
