+++ src/multiboot.h
@@ -16,30 +16,30 @@
 #define MULTIBOOT_FLAG_VBE     0x400
 
 struct multiboot {
-    uint32 flags;
-    uint32 mem_lower;
-    uint32 mem_upper;
-    uint32 boot_device;
-    uint32 cmdline;
-    uint32 mods_count;
-    uint32 mods_addr;
-    uint32 num;
-    uint32 size;
-    uint32 addr;
-    uint32 shndx;
-    uint32 mmap_length;
-    uint32 mmap_addr;
-    uint32 drives_length;
-    uint32 drives_addr;
-    uint32 config_table;
-    uint32 boot_loader_name;
-    uint32 apm_table;
-    uint32 vbe_control_info;
-    uint32 vbe_mode_info;
-    uint32 vbe_mode;
-    uint32 vbe_interface_seg;
-    uint32 vbe_interface_off;
-    uint32 vbe_interface_len;
+    size_t flags;
+    size_t mem_lower;
+    size_t mem_upper;
+    size_t boot_device;
+    size_t cmdline;
+    size_t mods_count;
+    size_t mods_addr;
+    size_t num;
+    size_t size;
+    size_t addr;
+    size_t shndx;
+    size_t mmap_length;
+    size_t mmap_addr;
+    size_t drives_length;
+    size_t drives_addr;
+    size_t config_table;
+    size_t boot_loader_name;
+    size_t apm_table;
+    size_t vbe_control_info;
+    size_t vbe_mode_info;
+    size_t vbe_mode;
+    size_t vbe_interface_seg;
+    size_t vbe_interface_off;
+    size_t vbe_interface_len;
 }  __attribute__((packed));
 
 typedef struct multiboot_header multiboot_header_t;
